refactor click commands into groups - local, remote, env - drop dev and stack hacks
