+++ hokusai/cli/__init__.py
@@ -1,6 +1,4 @@
-from hokusai.cli.base import cli
-from hokusai.cli.dev import dev
+from hokusai.cli.base import base
 from hokusai.cli.local import local
 from hokusai.cli.remote import remote
 from hokusai.cli.remote.env import env
-from hokusai.cli.remote.stack import stack
